Problem with dithered pixels in lines fixed
Due to problems with stroke() of lines using dithered pixels the fill
was looking bad so used a custom line drawing algorithm which might not
look the best for the stars so can have solid fill.  I suggest not to
use the stars and use a custom outline image as this turns out to be
much better.